+++ src/main/java/svenhjol/charmony/api/chest_puzzles/ChestPuzzleMenu.java
@@ -1,0 +1,9 @@
+package svenhjol.charmony.api.chest_puzzles;
+
+import svenhjol.charmony.api.stone_chests.StoneChestMaterial;
+
+public interface ChestPuzzleMenu {
+    StoneChestMaterial getMaterial();
+
+    ChestPuzzleType puzzleType();
+}