+++ src/main/java/svenhjol/charmony/api/chest_puzzles/ChestPuzzleType.java
@@ -1,0 +1,7 @@
+package svenhjol.charmony.api.chest_puzzles;
+
+public enum ChestPuzzleType {
+    Other,
+    Item,
+    Time
+}