+++ src/main/java/svenhjol/charmony/api/secret_chests/SecretChestPuzzleMenuProvider.java
@@ -1,13 +1,13 @@
-package svenhjol.charmony.api.stone_chests;
+package svenhjol.charmony.api.secret_chests;
 
-import net.minecraft.world.inventory.AbstractContainerMenu;
+import svenhjol.charmony.api.chest_puzzles.ChestPuzzleMenu;
 
 import java.util.Optional;
 
 /**
  * Add a custom menu that can be used when the player attempts to open a locked stone chest.
  */
-public interface StoneChestLockMenuProvider {
+public interface SecretChestPuzzleMenuProvider {
     /**
      * ID of the custom menu.
      * This will be used as a map key for this provider.
@@ -18,5 +18,5 @@ public interface StoneChestLockMenuProvider {
      * Custom menu to open when the chest is interacted with by the player.
      * Return an empty optional if something goes wrong in the processing of the custom menu.
      */
-    Optional<AbstractContainerMenu> getMenuProvider(StoneChestLockMenuData menuData);
+    Optional<ChestPuzzleMenu> getMenuProvider(SecretChestPuzzleMenuData menuData);
 }