+++ src/main/java/svenhjol/charmony/stone_chests/common/features/chest_puzzles/MenuProviders.java
@@ -1,7 +1,7 @@
 package svenhjol.charmony.stone_chests.common.features.chest_puzzles;
 
 import svenhjol.charmony.api.Api;
-import svenhjol.charmony.api.stone_chests.StoneChestLockMenuProvider;
+import svenhjol.charmony.api.secret_chests.SecretChestPuzzleMenuProvider;
 import svenhjol.charmony.core.base.Setup;
 import svenhjol.charmony.stone_chests.common.features.chest_puzzles.puzzles.*;
 import svenhjol.charmony.stone_chests.common.features.chest_puzzles.puzzles.EnderItemPuzzleProvider;
@@ -14,7 +14,7 @@ public class MenuProviders extends Setup<ChestPuzzles> {
     public MenuProviders(ChestPuzzles feature) {
         super(feature);
 
-        List<? extends StoneChestLockMenuProvider> providers = List.of(
+        List<? extends SecretChestPuzzleMenuProvider> providers = List.of(
             new ClockPuzzleMenuProvider(),
             new EnchantedBookPuzzleMenuProvider(),
             new EnchantedItemPuzzleMenuProvider(),