+++ src/main/java/svenhjol/charmony/stone_chests/common/features/chest_puzzles/PuzzleMenu.java
@@ -1,7 +1,0 @@
-package svenhjol.charmony.stone_chests.common.features.chest_puzzles;
-
-import svenhjol.charmony.api.stone_chests.StoneChestMaterial;
-
-public interface PuzzleMenu {
-    StoneChestMaterial getMaterial();
-}