+++ src/main/java/svenhjol/charmony/stone_chests/common/features/chest_puzzles/puzzles/BaseItemPuzzleMenuProvider.java
@@ -1,19 +1,19 @@
 package svenhjol.charmony.stone_chests.common.features.chest_puzzles.puzzles;
 
 import net.minecraft.Util;
-import net.minecraft.world.inventory.AbstractContainerMenu;
 import svenhjol.charmony.api.chest_puzzles.ItemPuzzleRequirement;
-import svenhjol.charmony.api.stone_chests.StoneChestLockMenuData;
-import svenhjol.charmony.api.stone_chests.StoneChestLockMenuProvider;
+import svenhjol.charmony.api.secret_chests.SecretChestPuzzleMenuData;
+import svenhjol.charmony.api.secret_chests.SecretChestPuzzleMenuProvider;
+import svenhjol.charmony.api.chest_puzzles.ChestPuzzleMenu;
 import svenhjol.charmony.stone_chests.common.features.chest_puzzles.menus.ItemPuzzleMenu;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public abstract class BaseItemPuzzleMenuProvider implements StoneChestLockMenuProvider {
+public abstract class BaseItemPuzzleMenuProvider implements SecretChestPuzzleMenuProvider {
     @Override
-    public Optional<AbstractContainerMenu> getMenuProvider(StoneChestLockMenuData menuData) {
+    public Optional<ChestPuzzleMenu> getMenuProvider(SecretChestPuzzleMenuData menuData) {
         var requirements = new ArrayList<>(getRequirements());
         if (requirements.isEmpty()) return Optional.empty();