+++ src/main/java/svenhjol/charmony/stone_chests/common/features/chest_puzzles/puzzles/EnderItemPuzzleProvider.java
@@ -1,12 +1,12 @@
 package svenhjol.charmony.stone_chests.common.features.chest_puzzles.puzzles;
 
 import svenhjol.charmony.api.chest_puzzles.ItemPuzzleRequirement;
-import svenhjol.charmony.api.stone_chests.StoneChestLockMenuProvider;
+import svenhjol.charmony.api.secret_chests.SecretChestPuzzleMenuProvider;
 import svenhjol.charmony.stone_chests.common.features.chest_puzzles.Tags;
 
 import java.util.List;
 
-public class EnderItemPuzzleProvider extends BaseItemPuzzleMenuProvider implements StoneChestLockMenuProvider {
+public class EnderItemPuzzleProvider extends BaseItemPuzzleMenuProvider implements SecretChestPuzzleMenuProvider {
     public static final String ID = "ender_item_puzzle";
 
     @Override